[TEST] Reorganice un poco mejor el codigo de pruebas
